test(10C): 各ファイルに対応したテストへ分割\n\n- BattleCalculator.test.ts: 条件評価/潜在マージのケースを追加\n- ComboBoostManager.test.ts: 各種ブースト適用と消費の検証\n- BattleActionExecutor.test.ts: コンボ登録→次スキル消費の統合テスト\n- 旧: SkillFlexibility.extra.test.ts を削除
